+++ app/Home/layout.js
@@ -1,32 +1,13 @@
-import { Montserrat } from "next/font/google";
-import "./globals.css";
-import { Footer } from "./components/footer";
-import { Navbar } from "./components/navbar";
-import { Toaster } from "@/components/ui/sonner";
-import { Providers } from "@/components/ui/providers";
+import { Footer } from "../components/homecomponents/footer";
+import { Navbar } from "../components/homecomponents/navbar";
 
-const montserrat = Montserrat({
-  subsets: ["latin"],
-  weight: ["200", "400", "500", "700"],
-});
-
-export const metadata = {
-  title: "+IDEAS",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
+export default function HomeLayout({ children }) {
   return (
-    <html lang="en">
-      <body className={`${montserrat.className} antialiased`}>
-        <Providers>
-          <Navbar />
-          {children}
-          <Toaster />
-          <Footer />
-        </Providers>
-      </body>
-    </html>
+    <>
+      <Navbar />
+      {children}
+      <Footer />
+    </>
   );
 }
 //minusculas
